Add preliminary live-migration support (e.g. for Xen for FC6

+++ rgmanager/src/daemons/rg_queue.c
@@ -26,7 +26,7 @@
 int
 _rq_queue_request(request_t **queue, char *name, uint32_t request,
 		 uint32_t err, uint32_t oldreq, msgctx_t *ctx, time_t when,
-		 uint64_t target, uint32_t arg0, uint32_t arg1, char *file,
+		 uint32_t target, uint32_t arg0, uint32_t arg1, char *file,
 		 int line)
 {
 	request_t *req;
